fix: cargo warning in `decode_strings`
Previously what the function was doing was a little esoteric keeping and
updating two different references to `data`. This resulted in a compile
warning that `cargo clippy` incorrectly fixed.

This PR normalizes a bit to avoid the two references and use only read
methods on the slice. Also includes unit tests which pass on main and
this branch.

cc @joaomoreno can you validate this works irl?