Remove panics and improve db updating to remove duplicates (#52)
* use single reader for decoding strings

* print all files in db when running with --parse

* add tests

* remove duplicate delete record entries

* add fixture

* fix duplicate removal

* do not keep CRC as part of header

* add --update debug flag

* rebase should work on paths, not strings

* run tests

* debug build

* revert debug build

* simplify

* fix: cargo warning in `decode_strings`

Previously what the function was doing was a little esoteric keeping and
updating two different references to `data`. This resulted in a compile
warning that `cargo clippy` incorrectly fixed.

This PR normalizes a bit to avoid the two references and use only read
methods on the slice. Also includes unit tests which pass on main and
this branch.

cc @joaomoreno can you validate this works irl?

* fix tests

---------

Co-authored-by: Connor Peet <connor@peet.io>